MedGuardian_Project/requirements.txt
# MedGuardian — Python dependencies
# Install with: pip install -r requirements.txt
#
# NOTE: pytesseract also requires the Tesseract OCR binary installed on
# your OS. See README.md for platform-specific instructions.

streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.20.0
pdfplumber>=0.11.0
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.2.0
python-dotenv>=1.0.0
